\name{SPLIT-package}
\alias{SPLIT-package}
\alias{SPLIT}
\docType{package}
\title{
\packageTitle{SPLIT}
}
\description{
\packageDescription{SPLIT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SPLIT}
\packageIndices{SPLIT}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{SPLIT}

Maintainer: \packageMaintainer{SPLIT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
