% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLIT_generate_coefficients.R
\name{SPLIT_generate_coefficients}
\alias{SPLIT_generate_coefficients}
\title{SPLIT Regression Modeling - Coefficients Generation}
\usage{
SPLIT_generate_coefficients(
  x,
  y,
  variables.split,
  intercept = TRUE,
  group.model = c("glmnet", "LS")[1],
  lambdas = NULL,
  alpha = 0
)
}
\arguments{
\item{x}{Design matrix.}

\item{y}{Response vector.}

\item{variables.split}{A vector with the split of the variables into groups as values.}

\item{intercept}{Boolean variable to determine if there is intercept (default is TRUE) or not.}

\item{group.model}{Model used for the groups. Must be one of "glmnet" or "LS".}

\item{lambdas}{The shinkrage parameters for the "glmnet" regularization. If NULL (default), optimal values are chosen.}

\item{alpha}{Elastic net mixing parameter. Should be between 0 (default) and 1.}
}
\value{
A vector with the regression coefficients for the split.
}
\description{
\code{SPLIT_generate_coefficients} generates the coefficients for a particular split of variables into groups.
}
\examples{
# Setting the parameters
p <- 6
n <- 30
n.test <- 5000
group.beta <- -3
beta <- c(rep(1, 2), rep(group.beta, p-2))
rho <- 0.1
r <- 0.9
SNR <- 3
# Creating the target matrix with "kernel" set to rho
target_cor <- function(r, p){
  Gamma <- diag(p)
  for(i in 1:(p-1)){
    for(j in (i+1):p){
      Gamma[i,j] <- Gamma[j,i] <- r^(abs(i-j))
    }
  }
  return(Gamma)
}
# AR Correlation Structure
Sigma.r <- target_cor(r, p)
Sigma.rho <- target_cor(rho, p)
sigma.epsilon <- as.numeric(sqrt((t(beta) \%*\% Sigma.rho \%*\% beta)/SNR))
# Simulate some data
x.train <- mvnfast::rmvn(30, mu=rep(0,p), sigma=Sigma.r)
y.train <- 1 + x.train \%*\% beta + rnorm(n=n, mean=0, sd=sigma.epsilon)
x.test <- mvnfast::rmvn(n.test, mu=rep(0,p), sigma=Sigma.rho)
y.test <- 1 + x.test \%*\% beta + rnorm(n.test, sd=sigma.epsilon)

# Generating the coefficients for a fixed split
SPLIT_generate_coefficients(x.train, y.train, variables.split=matrix(c(1,2,1,2,1,2), nrow=1))

}
\author{
Anthony-Alexander Christidis, \email{anthony.christidis@stat.ubc.ca}
}
